requirements.txt
Django==1.9.5
psycopg2==2.6.1
requests==2.10.0
django-extensions
djangorestframework
django-address==0.1.5
django-andablog==2.1.1
